rename scope_map.rs to filter.rs
